+++ flake.nix
@@ -12,7 +12,17 @@
       nixpkgs,
       flake-utils,
     }:
-    flake-utils.lib.eachDefaultSystem (
+    {
+      nixosModules = {
+        roam-activity = ./modules/nixos.nix;
+        default = self.nixosModules.roam-activity;
+      };
+      homeManagerModules = {
+        roam-activity = ./modules/home-manager.nix;
+        default = self.homeManagerModules.roam-activity;
+      };
+    }
+    // flake-utils.lib.eachDefaultSystem (
       system:
       let
         pkgs = nixpkgs.legacyPackages.${system};
@@ -20,6 +30,7 @@
       {
         packages = {
           roam = pkgs.callPackage ./default.nix { };
+          roam-activity = pkgs.callPackage ./pkgs/roam-activity.nix { };
           default = self.packages.${system}.roam;
         };
 
@@ -27,6 +38,10 @@
           roam = {
             type = "app";
             program = "${self.packages.${system}.roam}/bin/roam";
+          };
+          roam-activity = {
+            type = "app";
+            program = "${self.packages.${system}.roam-activity}/bin/roam-activity";
           };
           default = self.apps.${system}.roam;
         };