+++ shell.nix
@@ -4,4 +4,5 @@
 
 {
   roam = pkgs.callPackage ./default.nix { };
+  roam-activity = pkgs.callPackage ./pkgs/roam-activity.nix { };
 }